Add support for docker
This commit adds basic support for docker image creation. A Dockerfile
has been added along with a Makefile that contains targets to assist
with the docker build process.

Run `make docker-build` to build the docker image. After it is built
the image can be run by executing `docker run journalbeat`. Run
`make docker-tag` to build the image and tag the docker image with the
current git tag.