Fix shutdown when the output is unavailable
When the publishing queue is full and we need to stop the beat we will get stuck in the main loop of `beater.(*Journalbeat).Run` in `libbeat/publisher.(*client).PublishEvent`.
Moving closing the publisher to the `beater.(*Journalbeat).Stop` and unblocking return from  `journal.Follow` allows to shutdown gracefully in case the output is blocked for whatever reason.