Don't include sourcefiles in Docker release
Previously, running `make docker-build` would build journalbeat
in the same container that would be released. This commit creates
another step that builds journalbeat in one container, then copies
the built binary out of it so that it can be added to the image that
is released. The result is a much smaller image with none of the
source code hanging around.